refactor: move actual helpers into util folder
