+++ src/enhanceConfig.ts
@@ -2,7 +2,7 @@ import { mergeDeepLeft } from 'ramda';
 import { Config } from './config';
 import { Git } from './git/Git';
 import { GitErrorType } from './git/GitError';
-import { getFileContent } from './io/file-util';
+import { getFileContent } from './io/file';
 
 interface DefaultConfig {
   git: {
