+++ src/util/text.ts
